+++ example/lib/const/resource.dart
@@ -1,0 +1,9 @@
+/// generate by resouce_generator library, shouldn't edit.
+class R {
+    
+  /// ![preview](file:///Users/caijinglong/Documents/GitHub/flutter_image_compress/example/./img/auto-angle.jpg)
+  static const String IMG_AUTO_ANGLE_JPG = "img/auto-angle.jpg";
+    
+  /// ![preview](file:///Users/caijinglong/Documents/GitHub/flutter_image_compress/example/./img/img.jpg)
+  static const String IMG_IMG_JPG = "img/img.jpg";
+}